Refactor the app layout
The previous layout was a bit clunky and had a few ugly bugs when resizing was done. Whats more future changes indicate a need for a layout with more control over the footer and sidebar sections. This commit establishes a better overall layout including rearranging some of the components in both the DOM structure as well as the file structure.