Temporarily repair slider
The slider was assuming that the data would be for the current year,
however the texas data is for 2012. Strange side effects resolved by
adding the year into the Slider code. This will ultimately have to be
reworked as displaying the year strains the layout in the already tight
slider space.